Fix the Kotlin HCE service and its missing Android resources
Run 5 cleared the resource-linking failure and reached Kotlin compilation,
which surfaced the first real bugs in the HCE service — the code that makes
Vector 2 work and that had never once been compiled.

Kotlin: byteArrayOf(0x87, ...) in ppseResponse() and fciWithPdol(). 0x87 is
135, outside Kotlin's signed Byte range, so the literal is an Int and
byteArrayOf rejects it. Both now use 0x87.toByte(). Lines 61-63 already had
this right, which is why only these two failed — the same file was half
correct.

Resources: apduservice.xml referenced @string/swip_hce_description,
@string/swip_hce_aid_group and @drawable/swip_hce_banner, none of which
existed. Added strings.xml and a placeholder banner. The strings are worded for
where they actually appear — Android's Tap & pay settings, listed next to real
wallets, where a user has to be able to tell at a glance that SWIP cannot pay
for anything.

Also widened bootstrap's restore from res/xml to the whole res/ tree.
apduservice.xml depends on values/ and drawable/, so restoring only xml/ left
the service pointing at resources that were never put back. The copy is
additive, so flutter create's styles and mipmaps survive.

The banner is a placeholder Ink field with a gold bar; a vector drawable cannot
carry text, so the real one needs a PNG export from the brand generator. Noted
in the file.

+++ app/android/app/src/main/kotlin/in/swip/app/SwipListenService.kt
@@ -208,7 +208,7 @@ class SwipListenService : HostApduService() {
             // 61 <len> [ 4F <len> AID | 87 01 priority ]
             val aidBytes = aid.hexToBytes()
             val inner = byteArrayOf(0x4F, aidBytes.size.toByte()) + aidBytes +
-                byteArrayOf(0x87, 0x01, (idx + 1).toByte())
+                byteArrayOf(0x87.toByte(), 0x01, (idx + 1).toByte())
             (byteArrayOf(0x61, inner.size.toByte()) + inner).toList()
         }.toByteArray()
 
@@ -232,7 +232,7 @@ class SwipListenService : HostApduService() {
 
         val label = "SWIP".toByteArray(Charsets.US_ASCII)
         val proprietary = tlv(0x50, label) +          // Application Label
-            byteArrayOf(0x87, 0x01, 0x01) +            // Priority Indicator
+            byteArrayOf(0x87.toByte(), 0x01, 0x01) +            // Priority Indicator
             tlv(0x9F38, pdol)                          // PDOL
         val body = tlv(0x84, aid) + tlv(0xA5, proprietary)
         return tlv(0x6F, body)
